Implemented branding backend code; Implemented SkinManager; Fixed some rare bugs

+++ src/utils/playlist.h
@@ -80,7 +80,7 @@ public:
     virtual QString next();
 protected:
     int allLength;
-    int magic;
+    double magic;
     QDateTime minPlayTime;
 };
